refactor(plugin): split workbench publication helpers
