+++ plugins/codex-security/plugin-files.json
@@ -70,6 +70,7 @@
     "scripts/workbench_findings.py",
     "scripts/workbench_native_indexes.py",
     "scripts/workbench_progress.py",
+    "scripts/workbench_publication.py",
     "scripts/workbench_remediation.py",
     "scripts/workbench_saved_results.py",
     "scripts/workbench_scan_history.py",
