+++ plugins/codex-security/scripts/workbench_cli.py
@@ -247,15 +247,11 @@ def parse_args(description: str) -> argparse.Namespace:
     preserve_scan.add_argument("--claim-token")
     preserve_scan.add_argument("--coordinator-generation", type=positive_int)
 
-    recovery_help = (
-        "Validate and republish retained checkpoints for a failed, non-canceled scan."
-    )
+    recovery_help = "Validate and republish retained checkpoints for a failed, non-canceled scan."
     recover_scan = subparsers.add_parser(
         "recover-scan-results", help=recovery_help, description=recovery_help
     )
-    recover_scan.add_argument(
-        "--scan-id", required=True, help="ID of the stopped scan to recover."
-    )
+    recover_scan.add_argument("--scan-id", required=True, help="ID of the stopped scan to recover.")
 
     write_scan_draft = subparsers.add_parser("write-scan-draft")
     write_scan_draft.add_argument("--scan-id", required=True)
